trying to send diagnoses to UserDiagnoses and move the rest of the patient's info to Home
