+++ react-ts/src/components/UserDiagnoses.tsx
@@ -1,15 +1,11 @@
 // import userEvent from "@testing-library/user-event";
-import { User, userLocalStorageKey } from "../model/User";
-import { useState, useEffect } from "react";
+import { User } from "../model/User";
 
-export const UserDiagnoses = () => {
-  const [patient, setPatient] = useState<User>();
+interface Props {
+  diagnoses?: User;
+}
 
-  useEffect(() => {
-    const patientString = localStorage.getItem(userLocalStorageKey);
-    setPatient(patientString ? JSON.parse(patientString) : undefined);
-  }, []);
-
+export const UserDiagnoses = ({ diagnoses }: Props) => {
   /**
    * TODO
    *
@@ -18,31 +14,11 @@ export const UserDiagnoses = () => {
    * 2. Display list of user diagnoses (see Todo)
    */
 
-  return patient ? (
-    <div className="diagnose__list">
-      <div className="box">
-        <p className="box__title">Date of birth</p>
-        <p className="box__info">
-          <>{patient.dateOfBirth}</>
-        </p>
-      </div>
-      <div className="box">
-        <p className="box__title">Weight</p>
-        <p className="box__info">{patient.weight} kg</p>
-      </div>
-      <div className="box">
-        <p className="box__title">Height</p>
-        <p className="box__info">{patient.height} cm</p>
-      </div>
-      <div className="box">
-        <p className="box__title">Diagnoses</p>
-        <div className="user__diagnoses">
-          <p className="box__info">{patient.diagnoses[0]}</p>
-          <p className="box__info">{patient.diagnoses[1]}</p>
-        </div>
-      </div>
+  return (
+    <div className="user__diagnoses">
+      <ul>
+        <li className="box__info">{diagnoses}</li>
+      </ul>
     </div>
-  ) : (
-    <p className="patient__message">Patient not found.</p>
   );
 };
