+++ react-ts/src/pages/Home.tsx
@@ -1,8 +1,16 @@
 import { UserDiagnoses } from "../components/UserDiagnoses";
+import { User, userLocalStorageKey } from "../model/User";
 import "./Home.css";
 import "../global.css";
+import { useState, useEffect } from "react";
 
 export const Home = () => {
+  const [patient, setPatient] = useState<User>();
+
+  useEffect(() => {
+    const patientString = localStorage.getItem(userLocalStorageKey);
+    setPatient(patientString ? JSON.parse(patientString) : undefined);
+  }, []);
   /**
    * TODO
    *
@@ -11,10 +19,36 @@ export const Home = () => {
    * 3. Display name of the user in <h1> tag (see Todo)
    * 4. Send user's diagnoses to UserDiagnoses component
    */
-  return (
+
+  return patient ? (
     <div className="container home-container">
-      <p className="user__name">Todo: Display name of the user here</p>
-      <UserDiagnoses />
+      <p className="user__name">
+        {patient.firstName} {patient.lastName}
+      </p>
+      <div className="diagnose__list">
+        <div className="box">
+          <p className="box__title">Date of birth</p>
+          <p className="box__info">
+            <>{patient.dateOfBirth}</>
+          </p>
+        </div>
+        <div className="box">
+          <p className="box__title">Weight</p>
+          <p className="box__info">{patient.weight} kg</p>
+        </div>
+        <div className="box">
+          <p className="box__title">Height</p>
+          <p className="box__info">{patient.height} cm</p>
+        </div>
+        <div className="box">
+          <p className="box__title">Diagnoses</p>
+          {patient.diagnoses.map((diagnose) => (
+            <UserDiagnoses key={diagnose} diagnoses={diagnose} />
+          ))}
+        </div>
+      </div>
     </div>
+  ) : (
+    <p className="patient__message">Patient not found.</p>
   );
 };
